+++ lib/UI/listening_history_page.dart
@@ -8,8 +8,8 @@ import 'package:podboi/DataModels/listening_history.dart';
 import 'package:podboi/DataModels/song.dart';
 import 'package:podboi/Helpers/helpers.dart';
 import 'package:podboi/Shared/episode_display_widget.dart';
+import 'package:podboi/UI/player/mini_player.dart';
 import 'package:podboi/UI/podboi_loader.dart';
-import 'package:podboi/UI/player.dart';
 
 class ListeningHistoryView extends StatelessWidget {
   ListeningHistoryView({Key? key, required this.ref}) : super(key: key);
